+++ atividade/Questao_5/AppArea.java
@@ -2,14 +2,16 @@ import java.util.Scanner;
 
 public class AppArea {
     public static void main(String[] args) {
-        double raio;
+        float raio;
 
         Scanner scan = new Scanner(System.in);
         Circulo circulo = new Circulo();
 
         System.out.println("Entre com o valor do raio");
-        raio = scan.nextDouble();
+        raio = scan.nextFloat();
+
         circulo.setRaio(raio);
+        
         System.out.printf("A área é: %.2f\n", circulo.Area());
 
     }
